Messy File
Messy file that contains lots of notes for both the original format of the graph (3 groups with varied weathering, each with 2 columns based on depth) as well as the updated version (2 groups separated by depth that contain 3 columns each of weathered data). "Take 3" is the code that works for the new version. "Jamies work to calculate statistics" is the code that works for the original format. Contains notes from all of the different codes we have used to try and get the format correct (using our own research as well as GPT)